+++ Posh-ACME/Public/Set-PAServer.ps1
@@ -1,61 +1,104 @@
 function Set-PAServer {
     [CmdletBinding()]
     param(
-        [Parameter(Mandatory,Position=0,ValueFromPipeline,ValueFromPipelineByPropertyName)]
+        [Parameter(Position=0,ValueFromPipeline,ValueFromPipelineByPropertyName)]
         [ValidateScript({Test-ValidDirUrl $_ -ThrowOnFail})]
         [Alias('location')]
         [string]$DirectoryUrl,
         [Parameter(ValueFromPipelineByPropertyName)]
-        [switch]$SkipCertificateCheck
+        [switch]$SkipCertificateCheck,
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [switch]$DisableTelemetry
     )
 
-    # convert WellKnown names to their associated Url
-    if ($DirectoryUrl -notlike 'https://*') {
-        $DirectoryUrl = $script:WellKnownDirs.$DirectoryUrl
-    }
-    Write-Debug "Using DirectoryUrl $DirectoryUrl"
+    Process {
 
-    # make sure we're actually changing servers before doing anything else
-    if (!$script:Dir -or !$script:Dir.location -or $DirectoryUrl -ne $script:Dir.location) {
+        # make sure we have either a DirectoryUrl or an existing active server
+        if (-not $DirectoryUrl -and -not $script:Dir) {
+            try { throw "No DirectoryUrl specified and no active server. Please specify a DirectoryUrl." }
+            catch { $PSCmdlet.ThrowTerminatingError($_) }
+        }
 
-        # reset child object references
-        $script:Acct = $null
-        $script:AcctFolder = $null
-        $script:Order = $null
+        # convert WellKnown names to their associated Url
+        if ($DirectoryUrl -and $DirectoryUrl -notlike 'https://*') {
+            $DirectoryUrl = $script:WellKnownDirs.$DirectoryUrl
+            Write-Debug "Using DirectoryUrl $DirectoryUrl"
+        }
 
-        # deal with cert validation options between PS editions
-        Set-CertValidation $script:Dir.SkipCertificateCheck
+        # determine whether we're changing servers
+        if ($DirectoryUrl -and (-not $script:Dir -or $script:Dir.location -ne $DirectoryUrl)) {
+            $serverChange = $true
+        }
 
-        Update-PAServer $DirectoryUrl -SkipCertificateCheck:$SkipCertificateCheck.IsPresent
+        # use the active server if an explicit one wasn't specified
+        if (-not $DirectoryUrl) {
+            $DirectoryUrl = $script:Dir.location
+        }
+        $dirFolder = ConvertTo-DirFolder $DirectoryUrl
 
-        # save to disk
-        $DirectoryUrl | Out-File (Join-Path (Get-ConfigRoot) 'current-server.txt') -Force -EA Stop
+        # check for first use of this server
+        if (-not (Test-Path $dirFolder -PathType Container)) {
+            $firstUse = $true
+        }
+
+        # prep the server metadata
+        $dir = Get-PAServer $DirectoryUrl -Quiet
+        $updateParams = @{ DirectoryUrl = $DirectoryUrl }
+        if ($dir) {
+            Write-Debug "Loading existing server metadata."
+            $updateParams.DisableTelemetry = $dir.DisableTelemetry
+            $updateParams.SkipCertificateCheck = $dir.SkipCertificateCheck
+        } else {
+            $updateParams.DisableTelemetry = $false
+            $updateParams.SkipCertificateCheck = $false
+        }
+        if ('DisableTelemetry' -in $PSBoundParameters.Keys) {
+            Write-Debug "Setting DisableTelemetry value to $($DisableTelemetry.IsPresent)"
+            $updateParams.DisableTelemetry = $DisableTelemetry.IsPresent
+        }
+        if ('SkipCertificateCheck' -in $PSBoundParameters.Keys) {
+            Write-Debug "Setting SkipCertificateCheck value to $($SkipCertificateCheck.IsPresent)"
+            $updateParams.SkipCertificateCheck = $SkipCertificateCheck.IsPresent
+        }
+
+        if ($serverChange) {
+            # deal with cert validation options
+            Set-CertValidation $updateParams.SkipCertificateCheck
+        }
+
+        # update the server metadata
+        Update-PAServer @updateParams
+
+        if ($serverChange) {
+            # save as the new active server
+            $DirectoryUrl | Out-File (Join-Path (Get-ConfigRoot) 'current-server.txt') -Force -EA Stop
+        }
 
         # reload config from disk
         Import-PAConfig -Level 'Server'
 
-        # Show a link to the TOS if this server has no accounts associated with it yet.
-        if (!(Get-PAAccount -List)) {
+        # Show a link to the TOS if this is the server's first use.
+        if ($firstUse) {
             Write-Host "Please review the Terms of Service here: $($script:Dir.meta.termsOfService)"
         }
     }
 
 
-
-
-
     <#
     .SYNOPSIS
-        Set the current ACME server.
+        Set the current ACME server and its configuration.
 
     .DESCRIPTION
-        Use this function to switch between ACME servers. Switching also refreshes the cached data for the new server.
+        Use this function to switch between ACME servers or change configuration settings for a server. Switching also refreshes the cached data for the new server.
 
     .PARAMETER DirectoryUrl
-        Either the URL to an ACME server's "directory" endpoint or one of the supported short names. Currently supported short names include LE_PROD (LetsEncrypt Production v2) and LE_STAGE (LetsEncrypt Staging v2).
+        Either the URL to an ACME server's "directory" endpoint or one of the supported short names. Currently supported short names include LE_PROD (LetsEncrypt Production v2), LE_STAGE (LetsEncrypt Staging v2), BUYPASS_PROD (BuyPass.com Production), and BUYPASS_TEST (BuyPass.com Testing).
 
     .PARAMETER SkipCertificateCheck
         If specified, disable certificate validation while using this server. This should not be necessary except in development environments where you are connecting to a self-hosted ACME server.
+
+    .PARAMETER DisableTelemetry
+        If specified, telemetry data will not be sent to the Posh-ACME team for actions associated with this server. The telemetry data that gets sent by default includes Posh-ACME version, PowerShell version, and generic OS platform (Windows/Linux/MacOS).
 
     .EXAMPLE
         Set-PAServer LE_PROD